Feat/offer/add auth

+++ offer-ms/src/main/java/fr/polytech/ig5/mnm/offerms/controllers/ApplicationController.java
@@ -6,6 +6,7 @@ import fr.polytech.ig5.mnm.offerms.models.Application;
 import fr.polytech.ig5.mnm.offerms.models.Offer;
 import fr.polytech.ig5.mnm.offerms.services.ApplicationService;
 import fr.polytech.ig5.mnm.offerms.services.OfferService;
+import fr.polytech.ig5.mnm.offerms.utils.JwtUtils;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -25,6 +26,9 @@ public class ApplicationController {
     private ModelMapper modelMapper;
 
     @Autowired
+    JwtUtils jwtUtils;
+
+    @Autowired
     ApplicationService applicationService;
 
     @Autowired
@@ -35,15 +39,11 @@ public class ApplicationController {
     }
 
     @GetMapping("/applications/")
-    public ResponseEntity<Object> index() {
-        List <Application> applications = this.applicationService.findAll();
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(applications);
-    }
+    public ResponseEntity<Object> findByWorkerId(
+            @RequestHeader (name="Authorization") String bearerToken) {
 
-    @GetMapping("/applications/findByWorkerId/{workerId}")
-    public ResponseEntity<Object> findByWorkerId(@PathVariable("workerId") UUID workerId) {
+        UUID workerId = jwtUtils.extractUUIDFromJWT("workerId", bearerToken);
+
         List <Application> applications = this.applicationService.findByWorkerId(workerId);
         return ResponseEntity
                 .status(HttpStatus.OK)
@@ -51,54 +51,179 @@ public class ApplicationController {
     }
 
     @GetMapping("/applications/findByOfferId/{offerId}")
-    public ResponseEntity<Object> findByOfferId(@PathVariable("offerId") UUID offerId) {
-        Optional<Offer> offer = offerService.find(offerId);
+    public ResponseEntity<Object> findByOfferId(
+            @PathVariable("offerId") UUID offerId,
+            @RequestHeader (name="Authorization") String bearerToken) {
 
-        if(offer.isEmpty()){
+        UUID companyId = jwtUtils.extractUUIDFromJWT("companyId", bearerToken);
+        Optional<Offer> optionalOffer = offerService.find(offerId);
+
+        if(optionalOffer.isEmpty()){
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
                     .body("Offer not found");
         }
 
-        List <Application> applications = this.applicationService.findByOffer(offer.get());
+        Offer offer = optionalOffer.get();
+        if(!offer.getCompanyId().equals(companyId)){
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("Unauthorized");
+        }
+
+        List <Application> applications = this.applicationService.findByOffer(offer);
+
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(applications);
     }
 
     @GetMapping("/applications/{id}")
-    public ResponseEntity<Object> get(@PathVariable("id") UUID id) {
-        Optional<Application> application = this.applicationService.find(id);
+    public ResponseEntity<Object> get(
+            @PathVariable("id") UUID id,
+            @RequestHeader (name="Authorization") String bearerToken) {
 
-        if(application.isEmpty()){
+        Optional<Application> optionalApplication = this.applicationService.find(id);
+
+        if(optionalApplication.isEmpty()){
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
                     .body("Application not found");
         }
+
+        Application application = optionalApplication.get();
+
+        UUID workerId = jwtUtils.extractUUIDFromJWT("workerId", bearerToken);
+        UUID companyId = jwtUtils.extractUUIDFromJWT("companyId", bearerToken);
+
+        if(companyId != null) {
+            if(!application.getOffer().getCompanyId().equals(companyId)){
+                return ResponseEntity
+                        .status(HttpStatus.UNAUTHORIZED)
+                        .body("Unauthorized");
+            }
+        } else {
+            if(!application.getWorkerId().equals(workerId)){
+                return ResponseEntity
+                        .status(HttpStatus.UNAUTHORIZED)
+                        .body("Unauthorized");
+            }
+        }
+
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(application);
     }
 
-    @PostMapping("offers/{offerId}/applications")
-    public ResponseEntity<Object> create(@PathVariable("offerId") UUID offerId, @Valid @RequestBody ApplicationCreateDTO applicationDTO) {
-        Optional<Offer> offer = offerService.find(offerId);
+    @PostMapping("/offers/{offerId}/applications")
+    public ResponseEntity<Object> create(
+            @PathVariable("offerId") UUID offerId,
+            @Valid @RequestBody ApplicationCreateDTO applicationDTO,
+            @RequestHeader (name="Authorization") String bearerToken) {
 
-        if(offer.isEmpty()){
+        UUID workerId = jwtUtils.extractUUIDFromJWT("workerId", bearerToken);
+
+        Optional<Offer> optionalOffer = offerService.find(offerId);
+
+        if(optionalOffer.isEmpty()){
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
                     .body("Offer not found");
         }
+        Offer offer = optionalOffer.get();
 
-        Application application = modelMapper.map(applicationDTO, Application.class);
-        application.setOffer(offer.get());
+        Optional<Application> optionalApplication = this.applicationService.findApplicationByWorkerIdAndOffer(workerId, offer);
 
-        Application applicationCreated = this.applicationService.create(application);
+        if(!optionalApplication.isEmpty()){
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body("Already applied for this offer");
+        }
+
+        Application applicationToCreate = modelMapper.map(applicationDTO, Application.class);
+        applicationToCreate.setOffer(offer);
+        applicationToCreate.setWorkerId(workerId);
+
+        Application applicationCreated = this.applicationService.create(applicationToCreate);
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
                 .body(applicationCreated);
     }
+
+    @PutMapping(value = "/applications/acceptByWorker/{id}")
+    public ResponseEntity<Object> acceptByWorker(
+            @PathVariable("id") UUID id,
+            @RequestHeader (name="Authorization") String bearerToken) {
+
+        UUID workerId = jwtUtils.extractUUIDFromJWT("workerId", bearerToken);
+
+        Optional<Application> optionalApplication = this.applicationService.find(id);
+        if(optionalApplication.isEmpty()){
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body("Application not found");
+        }
+        Application application = optionalApplication.get();
+
+        if(!application.getWorkerId().equals(workerId)){
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("Unauthorized");
+        }
+
+        if(!application.getIsValidatedByCompany()){
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body("Application not yet validated by the company");
+        }
+
+        // tout est bon
+        application.setIsValidatedByWorker(true);
+        Application updatedApplication =
+                applicationService.update(application);
+        // TODO : envoyer message kafka pour qu'un travail se créé
+
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(updatedApplication);
+
+    }
+
+    @PutMapping(value = "/applications/acceptByCompany/{id}")
+    public ResponseEntity<Object> acceptByCompany(
+            @PathVariable("id") UUID id,
+            @RequestHeader (name="Authorization") String bearerToken) {
+
+        UUID companyId = jwtUtils.extractUUIDFromJWT("companyId", bearerToken);
+
+        Optional<Application> optionalApplication = this.applicationService.find(id);
+        if(optionalApplication.isEmpty()){
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body("Application not found");
+        }
+        Application application = optionalApplication.get();
+
+        UUID companyIdFromOffer = application.getOffer().getCompanyId();
+        if(!companyIdFromOffer.equals(companyId)){
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("Unauthorized");
+        }
+
+        // tout est bon
+        application.setIsValidatedByCompany(true);
+        Application updatedApplication =
+                applicationService.update(application);
+
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(updatedApplication);
+
+    }
+
+    /* Plus d'utilisation pour le moment
 
     @PutMapping(value = "applications/{id}")
     public ResponseEntity<Object> update(@PathVariable("id") UUID id, @Valid @RequestBody ApplicationUpdateDTO applicationDTO) {
@@ -121,9 +246,28 @@ public class ApplicationController {
                 .status(HttpStatus.OK)
                 .body(updatedApplication);
     }
+     */
 
-    @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Object> deleteOffer(@PathVariable UUID id) {
+    @DeleteMapping(value = "/applications/{id}")
+    public ResponseEntity<Object> deleteOffer(
+            @PathVariable UUID id,
+            @RequestHeader (name="Authorization") String bearerToken) {
+
+        UUID workerId = jwtUtils.extractUUIDFromJWT("workerId", bearerToken);
+
+        Optional<Application> optionalApplication = this.applicationService.find(id);
+        if(optionalApplication.isEmpty()){
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body("Application not found");
+        }
+        Application application = optionalApplication.get();
+
+        if(!application.getWorkerId().equals(workerId)){
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("Unauthorized");
+        }
 
         var isRemoved = this.applicationService.delete(id);
